優化模組安裝/更新界面
+ 1.修正在windows下更新會有不必要的訊息出現的問題。
+ 2.優化模組安裝/更新界面
+ 3.將模組安裝更新獨立為檔案，以供其他模組開發引用

+++ sql/mysql.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `tad_adm` (
-  `uid` smallint(6) unsigned NOT NULL COMMENT 'uid',
+  `uid` mediumint(8) unsigned NOT NULL COMMENT 'uid',
   `email` varchar(255) NOT NULL COMMENT 'email',
   `result` varchar(255) NOT NULL COMMENT '檢查結果',
   `chk_date` datetime NOT NULL COMMENT '檢查日',
